Se agrega deletes

+++ app/src/main/java/com/rfz/appflotal/data/model/product/response/ProductResponse.kt
@@ -13,5 +13,5 @@ data class ProductResponse(
     val userId: Int,
 
     @SerializedName("fld_treadDepth")
-    val treadDepth: Int
+    val treadDepth: Float
 )
